+++ sdk/formrecognizer/ai-form-recognizer/src/prebuilt/models.ts
@@ -6,9 +6,11 @@ import { Document } from "../generated";
 import { DocumentField, toDocumentField } from "../models/fields";
 import { isAcronymic, uncapitalize } from "../util";
 import { BusinessCard, BusinessCardSchema } from "./modelSchemas/businessCard";
+import { HealthInsuranceCardUs, HealthInsuranceCardUsSchema } from "./modelSchemas/healthInsuranceCard";
 import { IdentityDocument, IdentityDocumentSchema } from "./modelSchemas/idDocument";
 import { Invoice, InvoiceSchema } from "./modelSchemas/invoice";
 import { Receipt, ReceiptSchema } from "./modelSchemas/receipt";
+import { VaccinationCard, VaccinationCardSchema } from "./modelSchemas/vaccinationCard";
 import { TaxUsW2, TaxUsW2Schema } from "./modelSchemas/w2";
 import { ArrayFieldSchema, FieldSchema, ModelSchema, ReifyPrebuiltSchema } from "./schema";
 
@@ -175,6 +177,8 @@ export const PrebuiltModels = {
   Invoice: createModelFromSchema(InvoiceSchema) as DocumentModel<Invoice>,
   Receipt: createModelFromSchema(ReceiptSchema) as DocumentModel<Receipt>,
   TaxUsW2: createModelFromSchema(TaxUsW2Schema) as DocumentModel<TaxUsW2>,
+  VaccinationCard: createModelFromSchema(VaccinationCardSchema) as DocumentModel<VaccinationCard>,
+  HealthInsuranceCard: createModelFromSchema(HealthInsuranceCardUsSchema) as DocumentModel<HealthInsuranceCardUs>,
 };
 
 // PrebuiltModels is defined `as const` so this assignment checks to make sure it has the appropriate type, and